Successfully call authenticated athlete and retrieve image profile URL

+++ app/src/main/java/com/company/athleteapiart/data/entities/OAuth2Entity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class OAuth2Entity(
     @PrimaryKey
-    val athleteId: Int,
+    val athleteId: Long,
     val receivedOn: Int,
     val accessToken: String,
     val refreshToken: String
